refactor: move run-queue-service into zero layer as zero-run-queue-service
Move the run queue service from the infra layer into the zero layer by
merging it with the existing zero-queue-service, renamed to
zero-run-queue-service. The queue's only callers are zero-layer code, so
this simplifies the architecture by eliminating cross-layer imports.

Key changes:
- Rename zero-queue-service.ts → zero-run-queue-service.ts
- Merge all run-queue-service.ts functions into the renamed file
- Break reverse dependency: dispatchCancelSideEffects now accepts a
  drain callback instead of queueDispatcher
- Update all import paths across the codebase
- Move test file to zero/__tests__/
- Delete src/lib/run/run-queue-service.ts

Closes #7654
